Refactor session management into modular components
